requirements.txt
numpy==2.0.2
scikit-learn==1.5.1
pyyaml==6.0.2
rich==13.7.1